Componets: ListOfShow, ListOfEpisodes,ShowCards are created
